+++ examples/Cassie/cassie_rbt_state_estimator.h
@@ -8,14 +8,12 @@
 
 #include "drake/systems/framework/leaf_system.h"
 
-#include "InEKF.h"
-#include "RobotState.h"
-#include "NoiseParams.h"
 #include "attic/multibody/rigidbody_utils.h"
 #include "systems/framework/output_vector.h"
 #include "systems/framework/timestamped_vector.h"
 #include "examples/Cassie/datatypes/cassie_out_t.h"
 #include "examples/Cassie/cassie_utils.h"
+#include "examples/Cassie/inekf/InEKF.h"
 
 namespace dairlib {
 namespace systems {
