+++ examples/Cassie/inekf/BUILD.bazel
@@ -1,0 +1,57 @@
+load("@drake//tools/lint:lint.bzl", "add_lint_tests")
+
+package(default_visibility = ["//visibility:public"])
+
+load(
+    "@drake//tools/skylark:drake_lcm.bzl",
+    "drake_lcm_cc_library",
+    "drake_lcm_java_library",
+    "drake_lcm_py_library",
+)
+load(
+    "@drake//tools/skylark:drake_py.bzl",
+    "drake_py_binary",
+    "drake_py_library",
+    "drake_py_unittest",
+)
+
+load(
+    "@drake//tools/skylark:pybind.bzl",
+    "drake_pybind_library",
+    "get_drake_py_installs",
+    "get_pybind_package_info",
+    "pybind_py_library",
+)
+
+cc_library(
+    name = "LieGroup",
+    srcs = ["LieGroup.cpp"],
+    hdrs = ["LieGroup.h"],
+    deps = ["@drake//:drake_shared_library"]
+)
+
+cc_library(
+    name = "RobotState",
+    srcs = ["RobotState.cpp"],
+    hdrs = ["RobotState.h"],
+    deps = [
+        ":LieGroup"
+    ]
+)
+
+cc_library(
+    name = "NoiseParams",
+    srcs = ["NoiseParams.cpp"],
+    hdrs = ["NoiseParams.h"],
+    deps = ["//examples/Cassie/inekf:LieGroup"]
+)
+
+cc_library(
+    name = "InEKF",
+    srcs = ["InEKF.cpp"],
+    hdrs = ["InEKF.h"],
+    deps = [
+        "//examples/Cassie/inekf:RobotState",
+        "//examples/Cassie/inekf:NoiseParams",
+    ]
+)
