+++ examples/Cassie/inekf/LieGroup.cpp
@@ -11,7 +11,7 @@
  *  @date   September 25, 2018
  **/
 
-#include "LieGroup.h"
+#include "examples/Cassie/inekf/LieGroup.h"
 
 namespace inekf {
