+++ examples/Cassie/inekf/RobotState.cpp
@@ -11,8 +11,8 @@
  *  @date   September 25, 2018
  **/
 
-#include "RobotState.h"
-#include "LieGroup.h"
+#include "examples/Cassie/inekf/RobotState.h"
+#include "examples/Cassie/inekf/LieGroup.h"
 
 namespace inekf {
